Add basic distributor
Can only propagate Gotify push notifications to already registered clients
so far, no client management implemented at all yet.

This is prepared for supporting different push providers, as the existing
ones seem to mostly target small self-hosted installations, nothing that
would be viable as a KDE-scale service.

So far push providers are configured via the config file, but this might
eventually be something for a KCM.