Rework connector test to use an out-of-process distributor
The shortcuts we took to make this work in-process don't work anymore
with the D-Bus interface changes in the new UnifiedPush spec.

So, do this properly and run the distributor in a separate process. That
makes things more complicated but also allows us to have a more realistic
test, especially for the delayed response messages.